+++ source/Public/Assert-IPAddress.ps1
@@ -1,6 +1,7 @@
 <#
     .SYNOPSIS
-        Asserts that the specified IP address is valid.
+        Asserts if the IP Address is valid and optionally validates
+        the IP Address against an Address Family.
 
     .DESCRIPTION
         Checks the IP address so that it is valid and do not conflict with address
@@ -16,21 +17,21 @@
     .EXAMPLE
         Assert-IPAddress -Address '127.0.0.1'
 
-        This will assert that the supplied address is a valid IPv4 address.
-        If it is not an exception will be thrown.
+        This will assert that the supplied address is a valid IPv4 address. If it
+        is not an exception will be thrown.
 
     .EXAMPLE
         Assert-IPAddress -Address 'fe80:ab04:30F5:002b::1'
 
-        This will assert that the supplied address is a valid IPv6 address.
-        If it is not an exception will be thrown.
+        This will assert that the supplied address is a valid IPv6 address. If it
+        is not an exception will be thrown.
 
     .EXAMPLE
         Assert-IPAddress -Address 'fe80:ab04:30F5:002b::1' -AddressFamily 'IPv6'
 
-        This will assert that address is valid and that it matches the
-        supplied address family. If the supplied address family does not match
-        the address an exception will be thrown.
+        This will assert that address is valid and that it matches the supplied
+        address family. If the supplied address family does not match the address
+        an exception will be thrown.
 #>
 function Assert-IPAddress
 {